+++ K6/api/tests/access-management/metadata/roles/get-resources.js
@@ -4,7 +4,7 @@ import { group } from "k6";
 import { RolesGetRoleResourcesQueryBuilder } from "../../../../../clients/access-management/metadata/roles/index.js";
 import { getOptions } from "../../../../../helpers.js";
 import { MetadataBuildingBlocks } from "../../../../building-blocks/access-management/metadata/index.js";
-import { getClients, setup } from "../common.js";
+import { getClients, setup } from "./common.js";
 
 export { setup };
